fix(apple): Fix menu display and activation (#5090)
When the menu was clicked, the system would show it without activating
the app. This meant that other actions launched from the menu like `Sign
in` could sometimes happen without their respective window coming to the
foreground.

To fix it, we need more control over how the menu is displayed. So we
add a click handler and open the menu manually. Note that to do this, we
need to ensure `statusItem.menu = nil` otherwise the system will default
to showing the menu for us, and not firing our handler.

fixes #4818