test(connlib): ensure effective DNS servers match (#5107)
DNS in connlib mainly follows the following rules:

1. If there are upstream DNS servers configured in the portal, use
those.
2. Without upstream DNS servers, use the ones configured on the system.
3. Assign one proxy IP per effective DNS server.

We can test these properties using the state machine tests by
introducing an event that connlib emits every time it updates its DNS
servers. Additionally, this unblocks #5083. To resolve DNS resources, we
need to know, which DNS server to send a packet too and thus, we need
knowledge of the assigned proxy IP in the tests.

Currently, all clients learn about these effective DNS servers via the
callbacks. Those are still fairly intertwined in the `Tunnel`. Once we
have the resources to untangle this, we should use the emitted event to
inform the clients about the new DNS servers. I've captured this tech
debt in #5106.