refactor(gui-client): de-dupe IPC server code, enable debug IPC service for Linux (#5074)
Refs #5022

The debug IPC service has been useful on Windows, and since there is
more refactoring to do, I want it on Linux too.

With this you can just do `sudo -E target/debug/firezone-client-ipc
debug-ipc-service` and it will launch an IPC service without messing
with systemd or installing anything. (Assuming the directory for the
socket is created)

```[tasklist]
### Before merging
- [ ] Check for regressions in Windows
- [ ] Check for regressions in Linux
```