chore(connlib): split allowed_ips into ipv4 and ipv6 in `ClientOnGateway`
To encode into the types that clients always have both ipv4 and ipv6 and
they are the only allowed source ips for any given client, we split
those into their specific fields in the `ClientOnGateway` struct and
update tests accordingly.

Furthermore, these will be used for the DNS refactor for ipv6-in-ipv4
and ipv4-in-ipv6 to set the source ip of outgoing packets, without
having to do additional routing or mappings. There will be more notes on
this on the corresponding PR.