chore(connlib): make device buffers smaller (#5145)
Since we expect a fixed MTU size, we can encode this in the size of the
buffers for the device, we will never read or write more than the 1280
MTU we expect.

Note that the `write_buf` needs an extra 16 bytes for the aead tag that
boringtun will copy over.