docs(gui-client): add instructions for auto-start (#5159)
Refs #5118

I did it for Windows, too. Putting the shortcut in `%APPDATA%` like this
only makes it auto-start for the current user. Twingate appears to put
it under `%PROGRAMDATA%`, which causes it to auto-start for all users.
We could also use the registry if creating shortcuts programmatically on
Windows is difficult.

I tested these on an Ubuntu 22.04 aarch64 VM and the Windows 11 dev
laptop.