ci: make output of `cargo test` smaller (#5219)
When the `tunnel_test` fails, it generates a lot of output because it
keeps printing the backtrace over and over. This makes it difficult to
access the input seed to the test. Copying this seed into a local
environment is the first step in debugging this, at which point the
backtrace can be enabled locally.

We also disable the `verbose: 1` config option. Users can always set
that using the `PROPTEST_VERBOSE` env variable.