test(snownet): migrate relays for both parties (#5302)
In production, the portal will signal disconnected relays to both the
client and the gateway. We should mimic this in the tests.

In #5283, we remove invalidation of candidates during the connection
setup which breaks this roaming test due to "unhandled messages". We
could ignore those but I'd prefer to setup the test such that we panic
on unhandled messages instead and thus, this seems to be the better fix.